# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8088/api
  description: Generated server url
paths:
...
  /master/tety:
    get:
      tags:
      - tety-controller
      operationId: findAllData_1
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    put:
      tags:
      - tety-controller
      operationId: updateData
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateTypeDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - tety-controller
      operationId: saveData_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateTypeDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/temi:
    get:
      tags:
      - temi-controller
      operationId: findAllData_3
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    put:
      tags:
      - temi-controller
      operationId: updateData_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateInsurancePremiDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - temi-controller
      operationId: saveData_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateInsurancePremiDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/arwg:
    get:
      tags:
      - arwg-controller
      operationId: findAllData_12
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    put:
      tags:
      - arwg-controller
      operationId: updateData_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AreaWorkGroupDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - arwg-controller
      operationId: saveData_13
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AreaWorkGroupDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
...
  /master/tewo:
    get:
      tags:
      - tewo-controller
      operationId: findAllData
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - tewo-controller
      operationId: saveData
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TewoDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - tewo-controller
      operationId: updateData_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TewoDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/testa:
    get:
      tags:
      - testa-controller
      operationId: findAllData_2
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - testa-controller
      operationId: saveData_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateServiceTaskDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - testa-controller
      operationId: updateData_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateServiceTaskDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/regp:
    get:
      tags:
      - regp-controller
      operationId: findAllData_4
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - regp-controller
      operationId: saveData_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionPlatDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - regp-controller
      operationId: updateData_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionPlatDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/provinsi:
    post:
      tags:
      - prov-controller
      operationId: saveData_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvinsiDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - prov-controller
      operationId: updateData_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvinsiDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/inty:
    get:
      tags:
      - inty-controller
      operationId: findAllData_5
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - inty-controller
      operationId: saveData_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsuranceTypeDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - inty-controller
      operationId: updateData_7
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsuranceTypeDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/ibme:
    get:
      tags:
      - ibme-controller
      operationId: findAllData_6
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - ibme-controller
      operationId: saveData_7
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IbmeDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - ibme-controller
      operationId: updateData_8
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IbmeDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/cities:
    get:
      tags:
      - city-controller
      operationId: findAllData_7
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - city-controller
      operationId: saveData_8
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitiesDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - city-controller
      operationId: updateData_9
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitiesDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/category:
    get:
      tags:
      - cate-controller
      operationId: findAllData_8
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - cate-controller
      operationId: saveData_9
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - cate-controller
      operationId: updateData_10
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/carm:
    get:
      tags:
      - carm-controller
      operationId: findAllData_9
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - carm-controller
      operationId: saveData_10
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarModelDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - carm-controller
      operationId: updateData_11
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarModelDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/carb:
    get:
      tags:
      - carb-controller
      operationId: findAllData_10
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - carb-controller
      operationId: saveData_11
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarBrandDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - carb-controller
      operationId: updateData_12
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarBrandDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/car-series:
    get:
      tags:
      - cars-controller
      operationId: findAllData_11
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    post:
      tags:
      - cars-controller
      operationId: saveData_12
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarSeriesDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    patch:
      tags:
      - cars-controller
      operationId: updateData_13
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarSeriesDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
...
  /master/tewo/{id}:
    get:
      tags:
      - tewo-controller
      operationId: findDataById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/tety/{id}:
    get:
      tags:
      - tety-controller
      operationId: findDataById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/testa/{id}:
    get:
      tags:
      - testa-controller
      operationId: findDataById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/temi/{id}:
    get:
      tags:
      - temi-controller
      operationId: findDataById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/regp/{id}:
    get:
      tags:
      - regp-controller
      operationId: findDataById_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/provinsi/{id}:
    get:
      tags:
      - prov-controller
      operationId: findDataById_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/inty/{id}:
    get:
      tags:
      - inty-controller
      operationId: findDataById_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/ibme/{id}:
    get:
      tags:
      - ibme-controller
      operationId: findDataById_7
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
    delete:
      tags:
      - ibme-controller
      operationId: destroyData
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/cities/{id}:
    get:
      tags:
      - city-controller
      operationId: findDataById_8
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/category/{id}:
    get:
      tags:
      - cate-controller
      operationId: findDataById_9
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/carm/{id}:
    get:
      tags:
      - carm-controller
      operationId: findDataById_10
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/carb/{id}:
    get:
      tags:
      - carb-controller
      operationId: findDataById_11
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/car-series/{id}:
    get:
      tags:
      - cars-controller
      operationId: findDataById_12
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /master/arwg/{id}:
    get:
      tags:
      - arwg-controller
      operationId: findDataById_13
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
...
components:
  schemas:
...
    TemplateTypeDto:
      type: object
      properties:
        tetyId:
          type: integer
          format: int64
        tetyName:
          type: string
          enum:
          - CHECK_FEASIBILITY
          - SERVICE_POLIS
          - SERVICE_CLAIM
          - ORDER_CREATE
          - ORDER_MODIFY
          - STATUS_OPEN
          - STATUS_PENDING
          - STATUS_CANCELLED
          - STATUS_CLOSED
        tetyGroup:
          type: string
          enum:
          - SERVICE
          - ORDER
          - STATUS
    TemplateInsurancePremiDto:
      type: object
      properties:
        temiId:
          type: integer
          format: int64
        temiName:
          maxLength: 256
          minLength: 0
          type: string
        temiRateMin:
          type: number
          format: double
        temiRateMax:
          type: number
          format: double
        temiNominal:
          type: number
          format: double
        temiType:
          pattern: ^(Category|Extend)$
          type: string
        temiZonesId:
          type: integer
          format: int64
        temiIntyName:
          maxLength: 25
          minLength: 0
          type: string
        temiCateId:
          type: integer
          format: int64
    AreaWorkGroupDto:
      required:
      - arwgCode
      type: object
      properties:
        arwgCode:
          type: string
        arwgDesc:
          maxLength: 55
          minLength: 0
          type: string
        arwgCityId:
          type: integer
          format: int32
    TewoDto:
      required:
      - tewoName
      type: object
      properties:
        tewoId:
          type: integer
          format: int64
        tewoName:
          maxLength: 55
          minLength: 0
          type: string
        tewoTestaId:
          type: integer
          format: int64
    TemplateServiceTaskDto:
      type: object
      properties:
        testaId:
          type: integer
          format: int64
        testaName:
          maxLength: 55
          minLength: 0
          type: string
        testaGroup:
          type: integer
          format: int64
    RegionPlatDto:
      required:
      - regpName
      type: object
      properties:
        regpName:
          maxLength: 3
          minLength: 0
          type: string
        regpDesc:
          maxLength: 35
          minLength: 0
          type: string
        regpProvId:
          type: integer
          format: int64
    ProvinsiDto:
      required:
      - provName
      type: object
      properties:
        provId:
          type: integer
          format: int64
        provName:
          maxLength: 85
          minLength: 0
          type: string
        prov_zones_id:
          type: integer
          format: int64
    InsuranceTypeDto:
      required:
      - intyName
      type: object
      properties:
        intyName:
          maxLength: 25
          minLength: 0
          type: string
        intyDesc:
          maxLength: 55
          minLength: 0
          type: string
    IbmeDto:
      type: object
      properties:
        ibmeId:
          type: integer
          format: int64
        ibmeDate:
          type: string
          format: date
        ibme_entityid_source:
          type: integer
          format: int32
        ibme_entityid_target:
          type: integer
          format: int32
        ibme_type:
          type: integer
          format: int32
        ibme_count:
          type: integer
          format: int32
    CitiesDto:
      required:
      - cityName
      type: object
      properties:
        cityId:
          type: integer
          format: int64
        cityName:
          maxLength: 85
          minLength: 0
          type: string
        cityProvId:
          type: integer
          format: int64
    CategoryDto:
      required:
      - cateName
      type: object
      properties:
        cateId:
          type: integer
          format: int64
        cateName:
          maxLength: 55
          minLength: 0
          type: string
    CarModelDto:
      required:
      - carmCarbId
      - carmName
      type: object
      properties:
        carmId:
          type: integer
          format: int64
        carmName:
          maxLength: 55
          minLength: 0
          type: string
        carmCarbId:
          type: integer
          format: int64
    CarBrandDto:
      required:
      - cabrName
      type: object
      properties:
        cabrID:
          type: integer
          format: int64
        cabrName:
          maxLength: 55
          minLength: 0
          type: string
    CarSeriesDto:
      required:
      - carsCarmId
      - carsName
      type: object
      properties:
        carsId:
          type: integer
          format: int64
        carsName:
          maxLength: 55
          minLength: 0
          type: string
        carsPassenger:
          type: integer
          format: int32
        carsCarmId:
          type: integer
          format: int64
...